Clear what CI warns about on every run
The one actual ##[warning] annotation on every green run was this:

  Node.js 20 is deprecated. The following actions target Node.js 20 but are
  being forced to run on Node.js 24: actions/checkout@v4, actions/setup-java@v4,
  gradle/actions/setup-gradle@v4

So the majors move to ones that target Node 24 natively — checkout v7,
setup-java v6, setup-gradle v6, and github-script v9 in the cleanup
workflow, which had the same problem. We use these with no exotic inputs, so
the jump is mechanical; both lanes were run locally before pushing.

While there, the ten Kotlin warnings printed in every build log. Four were
real dead code and are now gone: an always-true null check on the NPU rung I
added yesterday, an unnecessary safe call on a non-null CurriculumUnit, a
redundant toDouble() on an expression already Double, and an always-false
second null check in the ASR capture loop — the gate exists exactly when the
detector does, so one nullable now carries the pair. That last one is
behaviour-preserving: same objects, same order, same reset.

Three were deprecated Java APIs with real replacements: URL(String) and
URL(URL, String) in PackFetcher become URI parse-and-resolve (deprecated
since Java 20 because URL parses leniently — validating before opening a
connection is the point), and Locale(String) becomes forLanguageTag, which
is the BCP-47 form the profile already stores.

The rest are deliberate and now say so: the TRIM_MEMORY_* constants are
deprecated in API 34 but still delivered, and dropping onLowMemory would
lose the low-memory signal on the 8 GB 9a, which is the device the trim
policy exists for. Suppressed with that reason next to it.

Not fixed: "the Kotlin Gradle plugin was loaded multiple times". Hoisting it
to the root build with apply false does silence it and the jvmOnly lane
survives, but the full lane then fails resolving kotlin.android in :app
("already on the classpath with an unknown version") — the same class of
conflict the root build's comment documents for AGP. Fixing it properly
means restructuring plugin management across every module, which is not
worth doing inside a warning cleanup.

+++ app/src/main/kotlin/org/sisam/langtutor/ui/reward/RewardOverlay.kt
@@ -93,7 +93,7 @@ private fun DrawScope.drawBurst(burst: RewardBurst, t: Float, scratch: Path) {
 
         val x = originX +
             (cos(angle) * speed * t).toFloat() +
-            if (gentle) (unit * 0.12f * sin((t * 3.0 + phase * 6.28).toDouble())).toFloat() else 0f
+            if (gentle) (unit * 0.12f * sin(t * 3.0 + phase * 6.28)).toFloat() else 0f
         val y = originY +
             (sin(angle) * speed * t).toFloat() +
             GRAVITY * size.height * t * t
